refactor: move user general page
