+++ apps/nextjs/src/app/[locale]/manage/users/[userId]/general/_components/_delete-user-button.tsx
