+++ apps/nextjs/src/app/[locale]/manage/users/[userId]/security/_components/_change-password-form.tsx
