+++ apps/nextjs/src/components/user-avatar-menu.tsx
@@ -4,11 +4,8 @@ import type { ReactNode } from "react";
 import { useCallback, useEffect } from "react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
-import type {
-  SelectProps} from "@mantine/core";
 import {
   Center,
-  Group,
   Menu,
   Stack,
   Text,
@@ -80,6 +77,11 @@ export const UserAvatarMenu = ({ children }: UserAvatarMenuProps) => {
           {t("navigateDefaultBoard")}
         </Menu.Item>
         <Menu.Divider />
+
+        <Menu.Item p={0} closeMenuOnClick={false}>
+          <LanguageCombobox />
+        </Menu.Item>
+        <Menu.Divider />
         {Boolean(session.data) && (
           <>
             <Menu.Item
@@ -89,19 +91,16 @@ export const UserAvatarMenu = ({ children }: UserAvatarMenuProps) => {
             >
               {t("preferences")}
             </Menu.Item>
-            <Menu.Item p={0} closeMenuOnClick={false}>
-              <LanguageCombobox />
+
+            <Menu.Item
+              component={Link}
+              href="/manage"
+              leftSection={<IconTool size="1rem" />}
+            >
+              {t("management")}
             </Menu.Item>
           </>
         )}
-        <Menu.Divider />
-        <Menu.Item
-          component={Link}
-          href="/manage"
-          leftSection={<IconTool size="1rem" />}
-        >
-          {t("management")}
-        </Menu.Item>
         <Menu.Divider />
         {session.status === "authenticated" ? (
           <Menu.Item
